test(ledger): consolidate transaction and fee test files :test_tube:
Merge test files that shared fixtures across file boundaries with nothing
declaring the dependency:

- postgres/transaction: createbulk + updatebulk -> transaction_bulk_test.go;
  updatebulk consumed createbulk's bulkMockDB and generateTestTransaction(s).
  Drop the duplicate strPtr in favor of the existing generic ptr helper.
- http/in: rename transaction_test.go -> transaction_handler_behavior_test.go;
  it drives HTTP requests through a Huma app against the exported
  TransactionHandler methods, not the two helpers left in transaction_core.go.
- http/in: fold transaction_fee_tenant_test.go into
  transaction_fee_application_test.go, restoring 1:1 mapping with the two
  methods of transaction_fee_application.go.
- http/in: fold the fee helpers and harness-sanity integration files into
  transaction_fee_harness_integration_test.go; all three hang off feeHarness.
- Replace the hand-rolled containsSub/aliasContains scan with strings.Contains,
  collapse legsForAlias/debitLegsForAlias/feeCreditLegs into one legsFor, and
  move mustTxID to the harness file where its five callers can find it.
